feat: move outside alerter to utils
