+++ packages/dialect-react-ui/utils/useOutsideAlerter.ts
@@ -1,0 +1,30 @@
+import { RefObject, useEffect } from 'react';
+
+export function useOutsideAlerter(
+  wrapperRef: RefObject<HTMLDivElement>,
+  buttonRef: RefObject<HTMLButtonElement>,
+  close: CallableFunction
+) {
+  useEffect(() => {
+    /**
+     * Alert if clicked on outside of element
+     */
+    function handleClickOutside(event: MouseEvent) {
+      if (
+        wrapperRef.current &&
+        !wrapperRef.current.contains(event.target as Node) &&
+        buttonRef.current &&
+        !buttonRef.current.contains(event.target as Node)
+      ) {
+        close();
+      }
+    }
+
+    // Bind the event listener
+    document.addEventListener('mousedown', handleClickOutside);
+    return () => {
+      // Unbind the event listener on clean up
+      document.removeEventListener('mousedown', handleClickOutside);
+    };
+  }, [wrapperRef, buttonRef, close]);
+}
